move `_set_base_attributes` to util
